modify_cart-process

+++ app/controllers/orders_controller.rb
@@ -4,19 +4,27 @@ class OrdersController < ApplicationController
     @order = Order.new
     @cart_items = current_user.cart_items
    #お届け先住所取得
+    @address_array = []
+    @postal_array = []
+    current_user.addresses.each do |address|
+      @address_array << address.address
+      @postal_array << address.postal_code
+    end
+    #ラジオボタン選択用
+    @default_button_num = current_user.default_address - 1
+    #合計金額0だったら購入させない（url直打ちやthanksページからのブラウザバック対策）
+    subtotal = 0
+    @sum = 0
     @cart_items.each do |cart_item|
-      @address_array = []
-      @postal_array = []
-      current_user.addresses.each do |address|
-        @address_array << address.address
-        @postal_array << address.postal_code
-      end
       #在庫数を超えた数量がカートに入っていないかチェック(ビュー側でこの処理をしてjavascriptで書き換えられたら理想？)
       if cart_item.quantity > cart_item.product.inventry_status
         cart_item.quantity = cart_item.product.inventry_status
-        flash[:quantity_notice] = "カート内の数量が変更されました"
+        if cart_item.save
+          flash[:quantity_notice] = "カート内の数量が変更されました"
+        end
       end
-
+      subtotal = cart_item.product.price * cart_item.quantity
+      @sum += subtotal
     end
   end
 
@@ -24,13 +32,7 @@ class OrdersController < ApplicationController
   def create
     @order = Order.new(order_params)
     @order.user_id = current_user.id
-    # if @order.save!
-    #    redirect_to thanks_order_path(@order.id)
-    # else
-    #   flash[:notice] = "処理に失敗しました"
-    #   redirect_to new_order_path
-    # end
-    
+
     Order.transaction do
       require "date"
       @order.delivery_status = 1
@@ -55,38 +57,45 @@ class OrdersController < ApplicationController
           @order.postal_code = postal_array[2]
         else
           #仮にこう記述
+          # user_default_address = Address.find_by(: current_user.default_address)
+          # @order.delivery_address = user_default_address.address
+          # @order.postal_code = user_default_address.postal_code
           flash[:notice] = 'お届け先住所を選択してください'
+          redirect_to new_order_path
+          return
       end
       @order.save!
       #order_itemsテーブルに保存開始
       @cart_items = CartItem.where(user_id: @order.user_id)
       total_price = 0 #orderテーブル保存用
       @cart_items.each do |cart_item|
-        #order_itemsテーブルに保存開始
-        product = cart_item.product
+        unless cart_item.quantity == 0 then
+          #order_itemsテーブルに保存開始
+          product = cart_item.product
 
-        order_item = OrderItem.new
-        order_item.order_id = @order.id
-        order_item.o_artist_name = product.artist_name
-        order_item.o_product_name = product.product_name
-        order_item.o_product_image_name_id = product.product_image_name_id
-        order_item.o_price = product.price
-        order_item.o_greed_id = product.greed.id
-        order_item.o_quantity = cart_item.quantity
-        order_item.product_id = product.id
-        order_item.save!
-        #在庫・売り上げ
-        # binding.pry
-        product.inventry_status -= cart_item.quantity
-        product.sales_quantity += cart_item.quantity
-        product.save!
-        #orderのtotal_price用
-        total_price += product.price * cart_item.quantity
-        cart_item.destroy!
+          order_item = OrderItem.new
+          order_item.order_id = @order.id
+          order_item.o_artist_name = product.artist_name
+          order_item.o_product_name = product.product_name
+          order_item.o_product_image_name_id = product.product_image_name_id
+          order_item.o_price = product.price
+          order_item.o_greed_id = product.greed.id
+          order_item.o_quantity = cart_item.quantity
+          order_item.product_id = product.id
+          order_item.save!
+          #在庫・売り上げ
+          # binding.pry
+          product.inventry_status -= cart_item.quantity
+          product.sales_quantity += cart_item.quantity
+          product.save!
+          #orderのtotal_price用
+          total_price += product.price * cart_item.quantity
+          cart_item.destroy!
+        end
       end
       @order.total_price = total_price
       @order.save!
-      # @cart_items.destroy_all! #処理に失敗したので、一旦cart_item.destroyで代替
+      # @cart_items.destroy_all! #処理に失敗した + 数量0をカートに残すため使わない
     end
       redirect_to thanks_order_path(@order.id)
     rescue => e
